rename directory classification -> utils
